backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
scikit-learn==1.8.0
pandas
numpy
joblib
streamlit
google-generativeai
